# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/contracts:
    post:
      operationId: ContractsController_createContract
      summary: Create a new contract
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractDto'
      responses:
        '201':
          description: Contract created successfully
        '400':
          description: Bad request
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
    get:
      operationId: ContractsController_getGymContracts
      summary: Get contracts for gym
      parameters:
        - name: status
          required: false
          in: query
          schema:
            enum:
              - pending
              - active
              - expiring_soon
              - expired
              - cancelled
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: offset
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of contracts
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
  /api/v1/contracts/{id}/renew:
    post:
      operationId: ContractsController_renewContract
      summary: Renew an existing contract
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewContractDto'
      responses:
        '200':
          description: Contract renewed successfully
        '404':
          description: Contract not found
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
  /api/v1/contracts/{id}/freeze:
    post:
      operationId: ContractsController_freezeContract
      summary: Freeze a contract
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FreezeContractDto'
      responses:
        '200':
          description: Contract frozen successfully
        '404':
          description: Contract not found
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
  /api/v1/contracts/{id}/cancel:
    put:
      operationId: ContractsController_cancelContract
      summary: Cancel a contract
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Contract cancelled successfully
        '404':
          description: Contract not found
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
  /api/v1/contracts/{id}:
    get:
      operationId: ContractsController_getContract
      summary: Get contract details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Contract details
        '404':
          description: Contract not found
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
  /api/v1/contracts/client/{clientId}:
    get:
      operationId: ContractsController_getClientContracts
      summary: Get client contract history
      parameters:
        - name: clientId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: List of client contracts
      tags:
        - Contracts
      security:
        - gym-id: []
        - bearer: []
...
  /api/v1/check-ins:
    post:
      operationId: CheckInsController_createCheckIn
      summary: Create a check-in for a client
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckInDto'
      responses:
        '201':
          description: Check-in created successfully
        '400':
          description: Bad request
      tags:
        - Check-ins
      security:
        - gym-id: []
        - bearer: []
    get:
      operationId: CheckInsController_searchCheckIns
      summary: Search check-ins
      parameters:
        - name: clientId
          required: false
          in: query
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: startDate
          required: false
          in: query
          schema:
            example: '2024-01-01'
            type: string
        - name: endDate
          required: false
          in: query
          schema:
            example: '2024-01-31'
            type: string
        - name: limit
          required: false
          in: query
          schema:
            example: '10'
            type: string
        - name: offset
          required: false
          in: query
          schema:
            example: '0'
            type: string
      responses:
        '200':
          description: List of check-ins
      tags:
        - Check-ins
      security:
        - gym-id: []
        - bearer: []
  /api/v1/check-ins/{id}:
    get:
      operationId: CheckInsController_getCheckIn
      summary: Get check-in details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Check-in details
        '404':
          description: Check-in not found
      tags:
        - Check-ins
      security:
        - gym-id: []
        - bearer: []
    delete:
      operationId: CheckInsController_deleteCheckIn
      summary: Delete a check-in (today only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Check-in deleted successfully
        '400':
          description: Can only delete today's check-ins
        '404':
          description: Check-in not found
      tags:
        - Check-ins
      security:
        - gym-id: []
        - bearer: []
  /api/v1/check-ins/stats/{period}:
    get:
      operationId: CheckInsController_getGymCheckInStats
      summary: Get gym check-in statistics
      parameters:
        - name: period
          required: true
          in: query
          schema:
            enum:
              - day
              - week
              - month
            type: string
      responses:
        '200':
          description: Check-in statistics
      tags:
        - Check-ins
      security:
        - gym-id: []
        - bearer: []
  /api/v1/check-ins/client/{clientId}/history:
    get:
      operationId: CheckInsController_getClientCheckInHistory
      summary: Get client check-in history
      parameters:
        - name: clientId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Client check-in history
      tags:
        - Check-ins
      security:
        - gym-id: []
        - bearer: []
...
  /api/v1/dashboard/stats:
    get:
      operationId: DashboardController_getDashboardStats
      summary: Get dashboard statistics
      description: Get aggregated statistics for the gym dashboard including clients, contracts, revenue, and check-ins
      parameters: []
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStatsDto'
        '403':
          description: Forbidden
      tags:
        - Dashboard
      security:
        - gym-id: []
        - bearer: []
  /api/v1/dashboard/recent-activity:
    get:
      operationId: DashboardController_getRecentActivity
      summary: Get recent activity
      description: Get recent activities in the gym including check-ins, new clients, and contracts
      parameters:
        - name: limit
          required: false
          in: query
          description: Maximum number of activities to return
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: List of recent activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecentActivityDto'
        '403':
          description: Forbidden
      tags:
        - Dashboard
      security:
        - gym-id: []
        - bearer: []
  /api/v1/dashboard/expiring-contracts:
    get:
      operationId: DashboardController_getExpiringContracts
      summary: Get expiring contracts
      description: Get contracts that are expiring in the next 30 days
      parameters:
        - name: limit
          required: false
          in: query
          description: Maximum number of contracts to return
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: List of expiring contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpiringContractDto'
        '403':
          description: Forbidden
      tags:
        - Dashboard
      security:
        - gym-id: []
        - bearer: []
info:
  title: GymSpace API
  description: The GymSpace Gym Management System API
  version: '1.0'
  contact:
    name: GymSpace API Support
    email: support@gymspace.com
  license:
    name: MIT
...
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.gymspace.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    gym-id:
      type: apiKey
      in: header
      name: X-Gym-Id
  schemas:
...
    CreateContractDto:
      type: object
      properties:
        gymClientId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        gymMembershipPlanId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174001
        startDate:
          type: string
          example: '2024-01-01'
          description: Contract start date
        discountPercentage:
          type: number
          example: 0
          description: Discount percentage (0-100)
        customPrice:
          type: number
          example: 49.99
          description: Custom price override
        metadata:
          type: object
          example:
            paymentMethod: credit_card
            referredBy: John Doe
            notes: Special promotion
      required:
        - gymClientId
        - gymMembershipPlanId
        - startDate
    RenewContractDto:
      type: object
      properties:
        startDate:
          type: string
          example: '2024-02-01'
          description: New start date
        discountPercentage:
          type: number
          example: 0
          description: Discount percentage for renewal
        customPrice:
          type: number
          example: 49.99
          description: Custom price for renewal
        metadata:
          type: object
          example:
            renewalReason: loyalty
            notes: Renewal with loyalty discount
    FreezeContractDto:
      type: object
      properties:
        freezeStartDate:
          type: string
          example: '2024-01-15'
          description: Freeze start date
        freezeEndDate:
          type: string
          example: '2024-01-30'
          description: Freeze end date
        reason:
          type: string
          example: Medical reasons
          description: Reason for freezing
      required:
        - freezeStartDate
        - freezeEndDate
...
    CreateCheckInDto:
      type: object
      properties:
        clientId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        notes:
          type: string
          example: Client feeling good today
      required:
        - clientId
...
    DashboardStatsDto:
      type: object
      properties:
        totalClients:
          type: number
          description: Total number of clients in the gym
          example: 156
        activeClients:
          type: number
          description: Number of active clients (with active contracts)
          example: 142
        totalContracts:
          type: number
          description: Total number of contracts
          example: 168
        activeContracts:
          type: number
          description: Number of active contracts
          example: 142
        monthlyRevenue:
          type: number
          description: Monthly revenue for the current month
          example: 4250
        todayCheckIns:
          type: number
          description: Number of check-ins today
          example: 23
        expiringContractsCount:
          type: number
          description: Number of contracts expiring in the next 30 days
          example: 8
        newClientsThisMonth:
          type: number
          description: Number of new clients registered this month
          example: 12
      required:
        - totalClients
        - activeClients
        - totalContracts
        - activeContracts
        - monthlyRevenue
        - todayCheckIns
        - expiringContractsCount
        - newClientsThisMonth
    RecentActivityDto:
      type: object
      properties:
        id:
          type: string
          description: Activity ID
          example: 123e4567-e89b-12d3-a456-426614174000
        type:
          type: string
          description: Type of activity
          enum:
            - check_in
            - new_client
            - new_contract
            - contract_expired
          example: check_in
        description:
          type: string
          description: Description of the activity
          example: Check-in registrado
        timestamp:
          type: string
          description: Timestamp of the activity
          example: '2024-01-15T10:30:00Z'
        clientName:
          type: string
          description: Name of the client related to the activity
          example: Juan Pérez
        clientId:
          type: string
          description: ID of the client related to the activity
          example: 123e4567-e89b-12d3-a456-426614174000
      required:
        - id
        - type
        - description
        - timestamp
    ExpiringContractDto:
      type: object
      properties:
        id:
          type: string
          description: Contract ID
          example: 123e4567-e89b-12d3-a456-426614174000
        clientName:
          type: string
          description: Client name
          example: Ana Martínez
        clientId:
          type: string
          description: Client ID
          example: 123e4567-e89b-12d3-a456-426614174000
        planName:
          type: string
          description: Membership plan name
          example: Plan Premium
        endDate:
          type: string
          description: Contract end date
          example: '2024-01-20T00:00:00Z'
        daysRemaining:
          type: number
          description: Days remaining until expiration
          example: 5
      required:
        - id
        - clientName
        - clientId
        - planName
        - endDate
        - daysRemaining